% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.HR}
\alias{df.HR}
\title{Meta-analyses exploring the efficacy of several interventions on a binary outcome.}
\format{
The dataset contains the following variables: \tabular{ll}{
  \strong{meta_review} \tab name of the first author of the meta-analysis.\cr
  \tab \cr
  \strong{factor} \tab name of the intervention studied.\cr
  \tab \cr
  \strong{author} \tab first study author of the RCTs.\cr
  \tab \cr
  \strong{year} \tab year of publication.\cr
  \tab \cr
  \strong{measure} \tab type of effect size (HR).\cr
  \tab \cr
  \strong{value} \tab HR value.\cr
  \tab \cr
  \strong{ci_lo} \tab lower bound of the 95\% confidence interval.\cr
  \tab \cr
  \strong{ci_up} \tab upper bound of the 95\% confidence interval.\cr
  \tab \cr
  \strong{n_cases} \tab number of cases.\cr
  \tab \cr
  \strong{n_controls} \tab number of controls.\cr
  \tab \cr
}
}
\source{
No source, the data are entirely fictitious
}
\usage{
df.HR
}
\description{
Fictitious dataset of four meta-analyses of RCTs assessing the efficacy of yoga, aerobic training,
resistance training and mindfulness on a numeric outcome
}
\keyword{datasets}
