% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.IRR}
\alias{df.IRR}
\title{Meta-analysis exploring adverse events of smoking.}
\format{
The dataset contains the following variables: \tabular{ll}{
  \strong{meta_review} \tab name of the first author of the meta-analysis.\cr
  \tab \cr
  \strong{factor} \tab name of the neurodevelopmental disorders on which the effect of the risk factor\cr
  \tab is studied\cr
  \tab \cr
  \strong{author} \tab first study author of the retrospective cohorts.\cr
  \tab \cr
  \strong{year} \tab year of publication.\cr
  \tab \cr
  \strong{measure} \tab type of effect size (IRR).\cr
  \tab \cr
  \strong{value} \tab IRR value.\cr
  \tab \cr
  \strong{ci_lo} \tab lower bound of the 95\% confidence interval.\cr
  \tab \cr
  \strong{ci_up} \tab upper bound of the 95\% confidence interval.\cr
  \tab \cr
  \strong{n_cases} \tab number of cases (sum of the number of cases in the exposed and non-exposed\cr
  \tab groups).\cr
  \tab \cr
  \strong{n_cases_exp} \tab number of cases in the exposed group.\cr
  \tab \cr
  \strong{n_cases_nexp} \tab number of cases in the non-exposed group.\cr
  \tab \cr
  \strong{time} \tab total person-time at risk (sum of the person-time at risk in the exposed and\cr
  \tab non-exposed groups).\cr
  \tab \cr
  \strong{time_exp} \tab person-time at risk in the exposed group.\cr
  \tab \cr
  \strong{time_nexp} \tab person-time at risk in the non-exposed group.\cr
  \tab \cr
}
}
\source{
No source, the data are entirely fictitious
}
\usage{
df.IRR
}
\description{
Fictitious dataset of a meta-analysis of prospective controlled cohorts assessing adverse effects of smoking.
}
\keyword{datasets}
