% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.MC}
\alias{df.MC}
\title{Meta-analyses exploring the efficacy of an intervention on a continuous outcome measured before and after the intervention.}
\format{
The dataset contains the following variables: \tabular{ll}{
  \strong{meta_review} \tab name of the first author of the meta-analysis.\cr
  \tab \cr
  \strong{factor} \tab name of the population studied.\cr
  \tab \cr
  \strong{author} \tab first study author of the individual studies.\cr
  \tab \cr
  \strong{year} \tab year of publication of the individual studies.\cr
  \tab \cr
  \strong{measure} \tab type of effect size (MC).\cr
  \tab \cr
  \strong{value} \tab MC value.\cr
  \tab \cr
  \strong{ci_lo} \tab lower bound of the 95\% confidence interval.\cr
  \tab \cr
  \strong{ci_up} \tab upper bound of the 95\% confidence interval.\cr
  \tab \cr
  \strong{n_cases} \tab number of cases.\cr
  \tab \cr
  \strong{n_controls} \tab number of controls.\cr
  \tab \cr
  \strong{mean_change_cases} \tab mean change score of patients in the experimental arm.\cr
  \tab \cr
  \strong{mean_change_controls} \tab mean change score of patients in the control arm.\cr
  \tab \cr
  \strong{sd_change_cases} \tab standard deviations of the change score of patients in the experimental arm.\cr
  \tab \cr
  \strong{sd_change_controls} \tab standard deviations of the change score of patients in the control arm.\cr
}
}
\source{
No source, the data are entirely fictitious
}
\usage{
df.MC
}
\description{
Fictitious dataset of three meta-analyses of RCTs assessing the efficacy of an intervention on a continuous outcome in 3 populations.
}
\keyword{datasets}
