% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.MD}
\alias{df.MD}
\title{Meta-analyses exploring the efficacy of surgical and pharmacological interventions.}
\format{
The dataset contains the following variables: \tabular{ll}{
  \strong{meta_review} \tab name of the first author of the meta-analysis.\cr
  \tab \cr
  \strong{factor} \tab name of the intervention studied.\cr
  \tab \cr
  \strong{author} \tab first study author of the individual studies.\cr
  \tab \cr
  \strong{year} \tab year of publication of the individual studies.\cr
  \tab \cr
  \strong{measure} \tab type of effect size (MD).\cr
  \tab \cr
  \strong{value} \tab MD value.\cr
  \tab \cr
  \strong{ci_lo} \tab lower bound of the 95\% confidence interval.\cr
  \tab \cr
  \strong{ci_up} \tab upper bound of the 95\% confidence interval.\cr
  \tab \cr
  \strong{mean_cases} \tab means of patients in the experimental arm.\cr
  \tab \cr
  \strong{mean_controls} \tab means of patients in the control arm.\cr
  \tab \cr
  \strong{sd_cases} \tab standard deviations of patients in the experimental arm.\cr
  \tab \cr
  \strong{sd_controls} \tab standard deviations of patients in the control arm.\cr
  \tab \cr
  \strong{n_cases} \tab number of patients in the experimental arm.\cr
  \tab \cr
  \strong{n_controls} \tab number of patients in the control arm.\cr
  \tab \cr
  \strong{rob} \tab risk of bias of the individual studies.\cr
  \tab \cr
  \strong{amstar} \tab AMSTAR score of the meta-analysis.\cr
}
}
\source{
No source, the data are entirely fictitious
}
\usage{
df.MD
}
\description{
Fictitious dataset of two meta-analyses of RCTs assessing the efficacy of surgical
 and pharmacological interventions on a numeric outcome.
}
\keyword{datasets}
