% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.OR.multi}
\alias{df.OR.multi}
\title{Meta-analysis of RCTs assessing different dietary interventions on a binary outcome.}
\format{
The dataset contains the following variables: \tabular{ll}{
  \strong{meta_review} \tab name of the first author of the meta-analysis.\cr
  \tab \cr
  \strong{factor} \tab name of the intervention studied.\cr
  \tab \cr
  \strong{author} \tab first study author of the individual studies\cr
  \tab \cr
  \strong{year} \tab year of publication of the individual studies.\cr
  \tab \cr
  \strong{measure} \tab type of effect size (OR).\cr
  \tab \cr
  \strong{value} \tab OR value.\cr
  \tab \cr
  \strong{ci_lo} \tab lower bound of the 95\% confidence interval.\cr
  \tab \cr
  \strong{ci_up} \tab upper bound of the 95\% confidence interval.\cr
  \tab \cr
  \strong{n_cases} \tab number of cases (sum of the number of cases in the exposed and\cr
  \tab non-exposed groups).\cr
  \tab \cr
  \strong{n_controls} \tab number of controls (sum of the number of controls in the exposed\cr
  \tab and non-exposed groups).\cr
  \tab \cr
  \strong{n_cases_exp} \tab number of cases in the exposed group.\cr
  \tab \cr
  \strong{n_controls_exp} \tab number of controls in the exposed group.\cr
  \tab \cr
  \strong{n_cases_nexp} \tab number of cases in the non-exposed group.\cr
  \tab \cr
  \strong{n_controls_nexp} \tab number of controls in the non-exposed group.\cr
  \tab \cr
  \strong{multiple_es} \tab indicates the reason of the presence for multiple effect sizes\cr
  \tab (due to multiple groups or outcomes) per study.\cr
}
}
\source{
No source, the data are entirely fictitious
}
\usage{
df.OR.multi
}
\description{
Fictitious dataset including meta-analyses with dependent effect sizes.
}
\keyword{datasets}
