% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.R}
\alias{df.R}
\title{Meta-analyses of correlational data}
\format{
The dataset contains the following variables: \tabular{ll}{
  \strong{meta_review} \tab name of the first author of the meta-analysis.\cr
  \tab \cr
  \strong{factor} \tab name of the factors.\cr
  \tab \cr
  \strong{author} \tab first study author of the individual studies.\cr
  \tab \cr
  \strong{year} \tab year of publication of the individual studies.\cr
  \tab \cr
  \strong{measure} \tab type of effect size (R).\cr
  \tab \cr
  \strong{value} \tab R value.\cr
  \tab \cr
  \strong{n_sample} \tab total number of individuals in the sample.\cr
}
}
\source{
No source, the data are entirely fictitious
}
\usage{
df.R
}
\description{
Fictitious dataset of four meta-analyses of cross-sectional studies exploring the association between
pre- or peri-pregnancy indicators and a numeric variable.
}
\keyword{datasets}
