% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_summary.R
\name{print.umbrella}
\alias{print.umbrella}
\title{Print a summary of an object of class \dQuote{umbrella}}
\usage{
\method{print}{umbrella}(x, ...)
}
\arguments{
\item{x}{an object of class \dQuote{umbrella}}

\item{...}{other arguments that can be passed to the function}
}
\value{
Implicitly calls the \code{\link{summary.umbrella}()} function and displays error or warning messages below the object returned.
This is useful when many factors are included in the review and that the results of the \code{\link{summary.umbrella}()} are not stored in an object.
}
\description{
Print a summary of an object of class \dQuote{umbrella}
}
\details{
Summary method for objects of class \dQuote{umbrella}.
}
\examples{
\donttest{
### print the results of an object of class umbrella
umbrella(df.OR.multi, mult.level = TRUE)
}
}
\seealso{
\code{\link{summary.umbrella}()}
}
