% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset.umbrella}
\alias{subset.umbrella}
\title{Create a subset of an object of class \dQuote{umbrella}}
\usage{
\method{subset}{umbrella}(x, subset, ...)
}
\arguments{
\item{x}{an object of class \dQuote{umbrella}.}

\item{subset}{logical expression indicating elements or rows to keep: missing values are taken as \code{FALSE}.}

\item{...}{other arguments that can be passed to the function}
}
\value{
Returns an object of class \dQuote{umbrella} with the results of some factors removed.
}
\description{
Create a subset of an object of class \dQuote{umbrella}
}
\examples{
### perform calculations required for an umbrella review
umb <- umbrella(df.OR)

### subset the results to factors "ADHD" & "dyslexia"
subset.umb <- subset.umbrella(umb, unique(df.OR$factor) \%in\% c("ADHD", "dyslexia"))

summary(subset.umb)
}
