% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.train}
\alias{df.train}
\title{Training dataset}
\format{
The dataset contains the following variables: \tabular{ll}{
  \strong{comment} \tab comments on studies.\cr
  \tab \cr
  \strong{risk_factor} \tab name of the intervention studied.\cr
  \tab \cr
  \strong{author_study} \tab first study author of the study\cr
  \tab \cr
  \strong{year_publication_study} \tab year of publication of the study.\cr
  \tab \cr
  \strong{type_of_effect_size} \tab type of effect size.\cr
  \tab \cr
  \strong{number_of_cases_exposed} \tab number of cases exposed\cr
  \tab \cr
  \strong{number_of_cases_non_exposed} \tab number of cases non-exposed\cr
  \tab \cr
  \strong{number_of_controls_exposed} \tab number of controls exposed\cr
  \tab \cr
  \strong{number_of_controls_non_exposed} \tab number of controls non-exposed\cr
  \tab \cr
  \strong{number_of_participants_exposed} \tab total number of participants exposed\cr
  \tab \cr
  \strong{number_of_participants_non_exposed} \tab total number of participants non-exposed\cr
  \tab \cr
  \strong{number_of_cases} \tab number of cases.\cr
  \tab \cr
  \strong{number_of_controls} \tab number of controls.\cr
  \tab \cr
  \strong{effect_size_value} \tab value of the effect size\cr
  \tab \cr
  \strong{low_bound_ci} \tab lower bound of the 95\% confidence interval.\cr
  \tab \cr
  \strong{up_bound_ci} \tab upper bound of the 95\% confidence interval.\cr
  \tab \cr
  \strong{time_disease_free} \tab time \cr
  \tab \cr
  \strong{mean_of_intervention_group} \tab mean of the intervention group\cr
  \tab \cr
  \strong{mean_of_control_group} \tab mean of the control group\cr
  \tab \cr
  \strong{sd_of_intervention_group} \tab sd of the intervention group\cr
  \tab \cr
  \strong{sd_of_control_group} \tab sd of the control group\cr
  \tab \cr
}
}
\source{
No source, the data are entirely fictitious
}
\usage{
df.train
}
\description{
This is a non-formatted dataset to illustrate how to obtain a well-formatted dataset.
}
\keyword{datasets}
